fix(meta): remove absent keys for likes and subscribers
Fixes #28

yt-dlp no longer seems to return information for the subscribers of a
channel, nor the like count of the video. While this is unfortunate, it
is just a minor detail.

This commit fixes this by just removing the display of likes and
subscribers from the frontend, and the return of these values from the
backend.